+++ sample-apps/react/egress-composite/src/components/UIDispatcher.tsx
@@ -1,11 +1,11 @@
 import { useCallStateHooks } from '@stream-io/video-react-sdk';
+
 import { useConfigurationContext } from '../ConfigurationContext';
 import { LayoutType, layoutMap } from './layouts';
-import { DominantSpeakerScreenShare } from './layouts/dominant-speaker';
-import { PaginatedGrid } from './layouts/grid';
+import { Spotlight } from './layouts/Spotlight';
 
-const DEFAULT_LAYOUT: LayoutType = 'grid';
-const DEFAULT_SCREENSHARE_LAYOUT: LayoutType = 'single_participant';
+const DEFAULT_LAYOUT: LayoutType = 'spotlight';
+const DEFAULT_SCREENSHARE_LAYOUT: LayoutType = 'spotlight';
 
 export const UIDispatcher = () => {
   const {
@@ -15,10 +15,9 @@ export const UIDispatcher = () => {
   const { useHasOngoingScreenShare } = useCallStateHooks();
   const hasScreenShare = useHasOngoingScreenShare();
 
-  const DefaultView = layoutMap[layout]?.[0] ?? PaginatedGrid;
+  const DefaultView = layoutMap[layout]?.[0] ?? Spotlight;
 
-  const ScreenShareView =
-    layoutMap[screenshare_layout]?.[1] ?? DominantSpeakerScreenShare;
+  const ScreenShareView = layoutMap[screenshare_layout]?.[1] ?? Spotlight;
 
   return hasScreenShare ? <ScreenShareView /> : <DefaultView />;
 };
