+++ sample-apps/react/egress-composite/src/components/layouts/DominantSpeaker/DominantSpeaker.tsx
