+++ sample-apps/react/egress-composite/src/components/layouts/DominantSpeaker/index.ts
