+++ sample-apps/react/egress-composite/src/components/layouts/PaginatedGrid/PaginatedGrid.scss
