+++ sample-apps/react/egress-composite/src/components/layouts/PaginatedGrid/index.ts
