+++ sample-apps/react/egress-composite/src/components/layouts/Spotlight/Spotlight.tsx
@@ -1,0 +1,33 @@
+import {
+  DefaultParticipantViewUI,
+  SpeakerLayout,
+} from '@stream-io/video-react-sdk';
+import { useConfigurationContext } from '../../../ConfigurationContext';
+
+import './Spotlight.scss';
+
+export const Spotlight = () => {
+  const {
+    options: { 'layout.spotlight.bar_position': barPosition = 'bottom' },
+  } = useConfigurationContext();
+
+  return (
+    <div className="spotlight">
+      <SpeakerLayout
+        participantsBarPosition={barPosition}
+        ParticipantViewUIBar={
+          <DefaultParticipantViewUI
+            indicatorsVisible={false}
+            showMenuButton={false}
+          />
+        }
+        ParticipantViewUISpotlight={
+          <DefaultParticipantViewUI
+            indicatorsVisible={false}
+            showMenuButton={false}
+          />
+        }
+      />
+    </div>
+  );
+};
