+++ sample-apps/react/egress-composite/src/components/layouts/Spotlight/index.ts
@@ -1,0 +1,1 @@
+export { Spotlight } from './Spotlight';
